+++ jslint4java-ant/src/test/resources/antunit/numeric-keys.js
@@ -1,2 +1,8 @@
-"use strict";
-var cfg = {'1': {'id': 1, 'fields': {}}, '2': {'id': 2, 'parentId': 1, 'fields': {'3': {'id': 3}}}};
+(function(){
+  "use strict";
+  var cfg = {'1': {'id': 1,
+                   'fields': {}},
+             '2': {'id': 2,
+                   'parentId': 1,
+                   'fields': {'3': {'id': 3}}}};
+}());